Added interval slider to pick the jam duration length.

+++ Main/JamChat/JamChat/HomeViewController.swift
@@ -95,14 +95,14 @@ class HomeViewController: UIViewController, UITableViewDelegate, UITableViewData
     }
     
     
-    func addNewJam(duration: Int, userIDs: [String]) {
+    func addNewJam(duration: Double, userIDs: [String]) {
         var jam: Jam!
         var jamLength = duration
         
         if userIDs.count == 0 {
             print("Can't create jam without users")
         } else {
-            jam = Jam(messageDuration: Double(jamLength), userIDs: userIDs)
+            jam = Jam(messageDuration: jamLength, userIDs: userIDs)
             jam.push { (success: Bool, error: NSError?) in
                 if let error = error {
                     print(error.localizedDescription)
